# apps/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/admin/accounts": {
      "get": {
        "operationId": "AdminController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedResponse"
                }
              }
            }
          },
          "400": {
            "description": "Request validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                },
                "example": {
                  "code": "VALIDATION_ERROR",
                  "message": "Request validation failed",
                  "errors": [],
                  "correlationId": "01HZZEXAMPLECORRELATION"
                }
              }
            }
          },
          "401": {
            "description": "Authentication is required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                },
                "example": {
                  "code": "UNAUTHENTICATED",
                  "message": "Authentication is required",
                  "errors": [],
                  "correlationId": "01HZZEXAMPLECORRELATION"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                },
                "example": {
                  "code": "NOT_FOUND",
                  "message": "Resource not found",
                  "errors": [],
                  "correlationId": "01HZZEXAMPLECORRELATION"
                }
              }
            }
          },
          "409": {
            "description": "The resource changed; reload and retry",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                },
                "example": {
                  "code": "VERSION_CONFLICT",
                  "message": "The resource changed; reload and retry",
                  "errors": [],
                  "correlationId": "01HZZEXAMPLECORRELATION"
                }
              }
            }
          },
          "422": {
            "description": "Manual classification is required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                },
                "example": {
                  "code": "MANUAL_CLASSIFICATION_REQUIRED",
                  "message": "Manual classification is required",
                  "errors": [],
                  "correlationId": "01HZZEXAMPLECORRELATION"
                }
              }
            }
          },
          "429": {
            "description": "Too many requests; try again later",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                },
                "example": {
                  "code": "RATE_LIMITED",
                  "message": "Too many requests; try again later",
                  "errors": [],
                  "correlationId": "01HZZEXAMPLECORRELATION"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "administration"
        ]
      }
    },
    "/api/v1/admin/accounts/{id}": {
      "get": {
        "operationId": "AdminController_detail",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "400": {
            "description": "Request validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                },
                "example": {
                  "code": "VALIDATION_ERROR",
                  "message": "Request validation failed",
                  "errors": [],
                  "correlationId": "01HZZEXAMPLECORRELATION"
                }
              }
            }
          },
          "401": {
            "description": "Authentication is required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                },
                "example": {
                  "code": "UNAUTHENTICATED",
                  "message": "Authentication is required",
                  "errors": [],
                  "correlationId": "01HZZEXAMPLECORRELATION"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                },
                "example": {
                  "code": "NOT_FOUND",
                  "message": "Resource not found",
                  "errors": [],
                  "correlationId": "01HZZEXAMPLECORRELATION"
                }
              }
            }
          },
          "409": {
            "description": "The resource changed; reload and retry",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                },
                "example": {
                  "code": "VERSION_CONFLICT",
                  "message": "The resource changed; reload and retry",
                  "errors": [],
                  "correlationId": "01HZZEXAMPLECORRELATION"
                }
              }
            }
          },
          "422": {
            "description": "Manual classification is required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                },
                "example": {
                  "code": "MANUAL_CLASSIFICATION_REQUIRED",
                  "message": "Manual classification is required",
                  "errors": [],
                  "correlationId": "01HZZEXAMPLECORRELATION"
                }
              }
            }
          },
          "429": {
            "description": "Too many requests; try again later",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                },
                "example": {
                  "code": "RATE_LIMITED",
                  "message": "Too many requests; try again later",
                  "errors": [],
                  "correlationId": "01HZZEXAMPLECORRELATION"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "administration"
        ]
      }
    },
    "/api/v1/admin/accounts/{id}/revoke-sessions": {
      "post": {
        "operationId": "AdminController_revokeSessions",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": true,
            "description": "Session-bound CSRF token",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "400": {
            "description": "Request validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                },
                "example": {
                  "code": "VALIDATION_ERROR",
                  "message": "Request validation failed",
                  "errors": [],
                  "correlationId": "01HZZEXAMPLECORRELATION"
                }
              }
            }
          },
          "401": {
            "description": "Authentication is required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                },
                "example": {
                  "code": "UNAUTHENTICATED",
                  "message": "Authentication is required",
                  "errors": [],
                  "correlationId": "01HZZEXAMPLECORRELATION"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                },
                "example": {
                  "code": "NOT_FOUND",
                  "message": "Resource not found",
                  "errors": [],
                  "correlationId": "01HZZEXAMPLECORRELATION"
                }
              }
            }
          },
          "409": {
            "description": "The resource changed; reload and retry",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                },
                "example": {
                  "code": "VERSION_CONFLICT",
                  "message": "The resource changed; reload and retry",
                  "errors": [],
                  "correlationId": "01HZZEXAMPLECORRELATION"
                }
              }
            }
          },
          "422": {
            "description": "Manual classification is required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                },
                "example": {
                  "code": "MANUAL_CLASSIFICATION_REQUIRED",
                  "message": "Manual classification is required",
                  "errors": [],
                  "correlationId": "01HZZEXAMPLECORRELATION"
                }
              }
            }
          },
          "429": {
            "description": "Too many requests; try again later",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                },
                "example": {
                  "code": "RATE_LIMITED",
                  "message": "Too many requests; try again later",
                  "errors": [],
                  "correlationId": "01HZZEXAMPLECORRELATION"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "administration"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MutationRequest"
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/accounts/{id}/reset-password": {
      "post": {
        "operationId": "AdminController_reset",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": true,
            "description": "Session-bound CSRF token",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "400": {
            "description": "Request validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                },
                "example": {
                  "code": "VALIDATION_ERROR",
                  "message": "Request validation failed",
                  "errors": [],
                  "correlationId": "01HZZEXAMPLECORRELATION"
                }
              }
            }
          },
          "401": {
            "description": "Authentication is required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                },
                "example": {
                  "code": "UNAUTHENTICATED",
                  "message": "Authentication is required",
                  "errors": [],
                  "correlationId": "01HZZEXAMPLECORRELATION"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                },
                "example": {
                  "code": "NOT_FOUND",
                  "message": "Resource not found",
                  "errors": [],
                  "correlationId": "01HZZEXAMPLECORRELATION"
                }
              }
            }
          },
          "409": {
            "description": "The resource changed; reload and retry",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                },
                "example": {
                  "code": "VERSION_CONFLICT",
                  "message": "The resource changed; reload and retry",
                  "errors": [],
                  "correlationId": "01HZZEXAMPLECORRELATION"
                }
              }
            }
          },
          "422": {
            "description": "Manual classification is required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                },
                "example": {
                  "code": "MANUAL_CLASSIFICATION_REQUIRED",
                  "message": "Manual classification is required",
                  "errors": [],
                  "correlationId": "01HZZEXAMPLECORRELATION"
                }
              }
            }
          },
          "429": {
            "description": "Too many requests; try again later",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                },
                "example": {
                  "code": "RATE_LIMITED",
                  "message": "Too many requests; try again later",
                  "errors": [],
                  "correlationId": "01HZZEXAMPLECORRELATION"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "administration"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MutationRequest"
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/accounts/{id}/activate": {
      "patch": {
        "operationId": "AdminController_activate",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": true,
            "description": "Session-bound CSRF token",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "400": {
            "description": "Request validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                },
                "example": {
                  "code": "VALIDATION_ERROR",
                  "message": "Request validation failed",
                  "errors": [],
                  "correlationId": "01HZZEXAMPLECORRELATION"
                }
              }
            }
          },
          "401": {
            "description": "Authentication is required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                },
                "example": {
                  "code": "UNAUTHENTICATED",
                  "message": "Authentication is required",
                  "errors": [],
                  "correlationId": "01HZZEXAMPLECORRELATION"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                },
                "example": {
                  "code": "NOT_FOUND",
                  "message": "Resource not found",
                  "errors": [],
                  "correlationId": "01HZZEXAMPLECORRELATION"
                }
              }
            }
          },
          "409": {
            "description": "The resource changed; reload and retry",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                },
                "example": {
                  "code": "VERSION_CONFLICT",
                  "message": "The resource changed; reload and retry",
                  "errors": [],
                  "correlationId": "01HZZEXAMPLECORRELATION"
                }
              }
            }
          },
          "422": {
            "description": "Manual classification is required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                },
                "example": {
                  "code": "MANUAL_CLASSIFICATION_REQUIRED",
                  "message": "Manual classification is required",
                  "errors": [],
                  "correlationId": "01HZZEXAMPLECORRELATION"
                }
              }
            }
          },
          "429": {
            "description": "Too many requests; try again later",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                },
                "example": {
                  "code": "RATE_LIMITED",
                  "message": "Too many requests; try again later",
                  "errors": [],
                  "correlationId": "01HZZEXAMPLECORRELATION"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "administration"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MutationRequest"
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/accounts/{id}/deactivate": {
      "patch": {
        "operationId": "AdminController_deactivate",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": true,
            "description": "Session-bound CSRF token",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "400": {
            "description": "Request validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                },
                "example": {
                  "code": "VALIDATION_ERROR",
                  "message": "Request validation failed",
                  "errors": [],
                  "correlationId": "01HZZEXAMPLECORRELATION"
                }
              }
            }
          },
          "401": {
            "description": "Authentication is required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                },
                "example": {
                  "code": "UNAUTHENTICATED",
                  "message": "Authentication is required",
                  "errors": [],
                  "correlationId": "01HZZEXAMPLECORRELATION"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                },
                "example": {
                  "code": "NOT_FOUND",
                  "message": "Resource not found",
                  "errors": [],
                  "correlationId": "01HZZEXAMPLECORRELATION"
                }
              }
            }
          },
          "409": {
            "description": "The resource changed; reload and retry",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                },
                "example": {
                  "code": "VERSION_CONFLICT",
                  "message": "The resource changed; reload and retry",
                  "errors": [],
                  "correlationId": "01HZZEXAMPLECORRELATION"
                }
              }
            }
          },
          "422": {
            "description": "Manual classification is required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                },
                "example": {
                  "code": "MANUAL_CLASSIFICATION_REQUIRED",
                  "message": "Manual classification is required",
                  "errors": [],
                  "correlationId": "01HZZEXAMPLECORRELATION"
                }
              }
            }
          },
          "429": {
            "description": "Too many requests; try again later",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                },
                "example": {
                  "code": "RATE_LIMITED",
                  "message": "Too many requests; try again later",
                  "errors": [],
                  "correlationId": "01HZZEXAMPLECORRELATION"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "administration"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MutationRequest"
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/employees/search": {
      "get": {
        "operationId": "AdminController_search",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "400": {
            "description": "Request validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                },
                "example": {
                  "code": "VALIDATION_ERROR",
                  "message": "Request validation failed",
                  "errors": [],
                  "correlationId": "01HZZEXAMPLECORRELATION"
                }
              }
            }
          },
          "401": {
            "description": "Authentication is required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                },
                "example": {
                  "code": "UNAUTHENTICATED",
                  "message": "Authentication is required",
                  "errors": [],
                  "correlationId": "01HZZEXAMPLECORRELATION"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                },
                "example": {
                  "code": "NOT_FOUND",
                  "message": "Resource not found",
                  "errors": [],
                  "correlationId": "01HZZEXAMPLECORRELATION"
                }
              }
            }
          },
          "409": {
            "description": "The resource changed; reload and retry",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                },
                "example": {
                  "code": "VERSION_CONFLICT",
                  "message": "The resource changed; reload and retry",
                  "errors": [],
                  "correlationId": "01HZZEXAMPLECORRELATION"
                }
              }
            }
          },
          "422": {
            "description": "Manual classification is required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                },
                "example": {
                  "code": "MANUAL_CLASSIFICATION_REQUIRED",
                  "message": "Manual classification is required",
                  "errors": [],
                  "correlationId": "01HZZEXAMPLECORRELATION"
                }
              }
            }
          },
          "429": {
            "description": "Too many requests; try again later",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                },
                "example": {
                  "code": "RATE_LIMITED",
                  "message": "Too many requests; try again later",
                  "errors": [],
                  "correlationId": "01HZZEXAMPLECORRELATION"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "administration"
        ]
      }
    },
...
  },
  "info": {
    "title": "CARE API",
    "description": "",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "care_session"
      }
    },
    "schemas": {
      "ErrorEnvelope": {
        "type": "object",
        "required": [
          "code",
          "message",
          "errors",
          "correlationId"
        ],
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FieldError"
            }
          },
          "correlationId": {
            "type": "string"
          },
          "meta": {
            "type": "object",
            "additionalProperties": true
          }
        }
      },
      "FieldError": {
        "type": "object",
        "required": [
          "field",
          "code",
          "message"
        ],
        "properties": {
          "field": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      },
      "MutationRequest": {
        "type": "object",
        "description": "Endpoint-specific request. Aggregate mutations require expectedVersion.",
        "properties": {
          "expectedVersion": {
            "type": "integer",
            "minimum": 1
          }
        },
        "additionalProperties": true
      },
      "SuccessResponse": {
        "type": "object",
        "additionalProperties": true
      },
      "PagedResponse": {
        "type": "object",
        "required": [
          "items",
          "nextCursor"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "type": "object",
              "additionalProperties": true
            }
          },
          "nextCursor": {
            "type": "string",
            "nullable": true,
            "description": "Signed opaque cursor"
          }
        }
      },
...
    }
  }
}